GST to Tax , New Template Format, Tax added in All Templates, Confirmation Added before any Deletion

+++ src/js/controller/customer.js
@@ -61,17 +61,20 @@ module.exports = [
     };
 
     $scope.deleteCustomer = (id) => {
-      CustomerModule.deleteCustomer(id, () => {
-        $scope.customer.find((o, i) => {
-          if (o) {
-            if (o.id === id) {
-              delete $scope.customer[i];
-              $scope.$apply();
-              return; // stop searching
+      if (confirm('Are you sure you want to delete this customer')) {
+        CustomerModule.deleteCustomer(id, () => {
+          $scope.customer.find((o, i) => {
+            if (o) {
+              if (o.id === id) {
+                delete $scope.customer[i];
+                $scope.$apply();
+                return; // stop searching
+              }
             }
-          }
+          });
         });
-      });
+      }
+
     };
 
     $scope.editCustomer = function (customer) {
